permissions for offer and order

+++ offer/api/views.py
@@ -1,16 +1,23 @@
+import django_filters
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 
 from offer.api.serializers import OfferDetailUrlSerializer, OfferGetSerializer, OfferCreateSerializer, DetailSerializer, SingleOfferGetSerializer, SingleOfferPatchSerializer
 from offer.models import Offer, OfferDetail
+from offer.api.permissions import IsBusinessUserToCreateOffer, IsOwnerOfOfferOrAdmin
 from userprofile.models import UserProfile
 from rest_framework.pagination import PageNumberPagination
 from rest_framework import generics
 from rest_framework import viewsets
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import filters
+# from rest_framework import filters
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django_filters import rest_framework as filters
+from rest_framework.permissions import IsAuthenticated, AllowAny
+
+
 
 
 
@@ -20,24 +27,24 @@ class LargeResultsSetPagination(PageNumberPagination):
     max_page_size = 10000
 
 
-class OfferView(APIView):
-  
 
-    
+class OfferView(APIView):
+   
     serializer_class = OfferGetSerializer
     pagination_class = LargeResultsSetPagination
-        
+    permission_classes = [IsBusinessUserToCreateOffer]
+    
     
     def get(self, request):
         queryset = Offer.objects.all()
-        
+        self.check_object_permissions(request, queryset)
         order_param = self.request.query_params.get('ordering', None)
         if order_param is not None and order_param is not '':
             queryset = queryset.order_by(order_param)
         
-        # creator_id_param = self.request.query_params.get('creator_id', None)
-        # if creator_id_param is not None and creator_id_param is not '':
-        #     queryset = queryset.filter(user=creator_id_param)
+        creator_id_param = self.request.query_params.get('creator_id', None)
+        if creator_id_param is not None and creator_id_param is not '':
+            queryset = queryset.filter(user=creator_id_param)
 
         min_price_param = self.request.query_params.get('min_price', None)
         if min_price_param is not None and min_price_param is not '':
@@ -51,25 +58,20 @@ class OfferView(APIView):
         if search_param is not None and search_param is not '':
             queryset = queryset.filter(Q(title__icontains=search_param) | Q(description__icontains=search_param))
             
-
-        
-
         pagination_class = LargeResultsSetPagination
         paginator = pagination_class()
         page = paginator.paginate_queryset(queryset, request, view=self)
         serializer = self.serializer_class(page, many=True, context={'request': request})
         
-        
-        # serializer = self.serializer_class(offer, many=True, context={'request': request})
         return paginator.get_paginated_response(serializer.data)
-        # return Response(serializer.data)
     
     
     
     def post(self, request):
-
+       
         business_user = UserProfile.objects.get(user=request.user)
-
+        self.check_object_permissions(request, business_user)
+        
         serializer = OfferCreateSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(user=business_user)
@@ -77,14 +79,17 @@ class OfferView(APIView):
         return Response(serializer.errors)
 
 class SingleOfferView(APIView):
+    permission_classes = [IsOwnerOfOfferOrAdmin]
    
     def get(self, request,pk):
         offer_detail = Offer.objects.get(pk=pk)
+        self.check_object_permissions(request, offer_detail)
         serializer = SingleOfferGetSerializer(offer_detail, context={'request': request})
         return Response(serializer.data)
     
-    def patch(self, request,pk):
+    def patch(self, request, pk):
         offer = Offer.objects.get(pk=pk)
+        self.check_object_permissions(request, offer)
         serializer = SingleOfferPatchSerializer(offer, data=request.data, partial=True, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -94,6 +99,7 @@ class SingleOfferView(APIView):
     
     def delete(self, request, pk):
         offer_instance = Offer.objects.get(pk=pk)
+        self.check_object_permissions(request, offer_instance)
         offer_instance.delete()
         return Response({})
   
@@ -101,11 +107,12 @@ class SingleOfferView(APIView):
 
 class OfferDetailsView(APIView):
     serializer_class = DetailSerializer
+    permission_classes = [IsOwnerOfOfferOrAdmin]
 
     def get(self, request,pk):
 
         offer_detail = OfferDetail.objects.get(pk=pk)
-        print(offer_detail)
+        self.check_object_permissions(request, offer_detail)
         serializer = self.serializer_class(offer_detail, context={'request': request})
      
         return Response(serializer.data)
